% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_all_photosynthetic_induction_parameters.R
\name{calculate_all_photosynthetic_induction_parameters}
\alias{calculate_all_photosynthetic_induction_parameters}
\title{A function that calculates all photosynthetic induction parameters
for all files.}
\usage{
calculate_all_photosynthetic_induction_parameters(manual_check = TRUE)
}
\arguments{
\item{manual_check}{A boolean that regulates if the fits are checked.}
}
\value{
Returns a dataframe with all the parameters.
}
\description{
A function that calculates all photosynthetic induction parameters
  for all files.
  The directory it reads is
  "input_directory_licorfiles/photosynthetic_induction_data/".
  For all the files it calculates all the parameters.
  It returns a dataframe with all the parameters.

  This makes it easy to drop files in the photosynthetic induction data
  directory and run just one line to fit all the data.
}
\examples{
mydata <- calculate_all_photosynthetic_induction_parameters()

}
\author{
Sam Loontjens
}
