% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_light_induction_parameters.R
\name{calculate_light_induction_parameters}
\alias{calculate_light_induction_parameters}
\title{A function that calculates the light incuction parameters.}
\usage{
calculate_light_induction_parameters(dataframe, min_stepsize = 1)
}
\arguments{
\item{dataframe}{A dataframe from which to calculate the light parameters.}
}
\value{
Returns a list of the lightinductionindex, startindex,
timezero, PAR1 and PAR2.
}
\description{
A function that calculates the light incuction parameters.
  It is usefull for photosynthetic induction data.
  It returns a list of the lightinductionindex, startindex,
  timezero, PAR1 and PAR2.
  The lightinductionindex describes the index where the
  light changes for the first time.
  The startindex is the index where the light is stable again
  after the lightinductionindex.
  The timezero is the time where the light changes for the first time.
  PAR1 is the stable light intensity in the beginning.
  PAR2 is the stable light intensity at the end.
  Note, only takes increases in light into account.
}
\examples{
list_of_light_parameters <- calculate_light_induction_parameters(mydata, 5)

}
\author{
Sam Loontjens
}
