% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directories.R
\name{create_directories}
\alias{create_directories}
\title{Create directories where you can drop files for analysis.}
\usage{
create_directories()
}
\value{
Nothing
}
\description{
Create directories where you can drop files for analysis.
  You can drop licorfiles in the
  input_directory_licorfiles/light_response_data for lightresponsecurves.
  You can drop licorfiles in the
  input_directory_licorfiles/photosynthetic_induction_data for
  induction curves.
}
\examples{
create_directories()

}
\author{
Sam Loontjens
}
