% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_parameters}
\alias{extract_parameters}
\title{A function that extracts desired parameters from a dataframe.}
\usage{
extract_parameters(dataframe, parameters)
}
\arguments{
\item{dataframe}{A dataframe to extract from}

\item{parameters}{A list of parameter names to extract}
}
\value{
Returns a dataframe with only the desired parameters.
}
\description{
A function that extracts desired parameters from a dataframe.
  It returns a dataframe with only the desired parameters.
}
\examples{
parameters_to_extract <- c("A, Ci, elapsed, Qin")
new_dataframe <- extract_parameters(old_dataframe, parameters_to_extract)

}
\author{
Sam Loontjens
}
