% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_A_Ci_curve.R
\name{fit_A_Ci_curve}
\alias{fit_A_Ci_curve}
\title{A function that fits A/Ci curve.}
\usage{
fit_A_Ci_curve(
  dataframe,
  O2 = 21,
  title = "A/Ci curve",
  subtitle = "",
  manual_check = TRUE,
  save_plot = FALSE,
  save_path = "output_directory_licorfiles/A_Ci_plots/"
)
}
\arguments{
\item{dataframe}{A dataframe from which to calculate the fit parameters.}

\item{title}{A string for the title of the plot}

\item{manual_check}{A boolean to regulate if you want to manual check the data}
}
\value{
Returns a list of the check state, A1, A2, deltaA, other constands and t50
}
\description{
A function
}
\examples{
  parameters <- fit_photosynthetic_induction_A_curve(mydata,
                                                     "data from today")

}
\author{
Sam Loontjens
}
