% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_photosynthetic_decay_A_curve.R
\name{fit_photosynthetic_decay_A_curve}
\alias{fit_photosynthetic_decay_A_curve}
\title{A function that fits A of a photosynthetic decay.}
\usage{
fit_photosynthetic_decay_A_curve(
  dataframe,
  mean_width = 50,
  fit_width = 0,
  title = "Photosynthetic decay A curve",
  subtitle = "",
  manual_check = TRUE,
  save_plot = FALSE,
  save_path = "output_directory_licorfiles/photosynthetic_decay_A_plots/"
)
}
\arguments{
\item{dataframe}{A dataframe from which to calculate the fit parameters.}

\item{mean_width}{An integer of how many points to use for the means.}

\item{fit_width}{An integer of how many points to use for the fit.
If fit_width is smaller than mean_width then it takes all points.}

\item{title}{A string for the title of the plot}

\item{manual_check}{A boolean to regulate if you want to manual check the data}
}
\value{
Returns a list of the check state, A1, A2, deltaA, other constands and t50
}
\description{
A function that fits A of a photosynthetic decay curve.
  It fits the data from the point where the light changes.
  It calculates A2, A3.
  It uses nls to fit the data.
  You can check if the fit is good.
  It returns a list with all the parameters it used to fit the line.
  These include A1, A2, deltaA, other constands and t50.
  If the model gave a warning or error then the parameter values are NA.

  It only accepts licor files that have every input parameter stable except
  changing the light once. Commonly made with a timed lamp autoprogram with
  only two steps.The dataframe has to contain A, elapsed and Qin.
}
\examples{
  parameters <- fit_photosynthetic_induction_A_curve(mydata,
                                                     "data from today")

}
\author{
Sam Loontjens
}
