% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_photosynthetic_induction_Ci_curve.R
\name{fit_photosynthetic_induction_Ci_curve}
\alias{fit_photosynthetic_induction_Ci_curve}
\title{A function that fits Ci of a photosynthetic induction.}
\usage{
fit_photosynthetic_induction_Ci_curve(
  dataframe,
  title = "PI Ci curve",
  mean_width = 50,
  fit_width = 0,
  manual_check = TRUE
)
}
\arguments{
\item{dataframe}{A dataframe from which to calculate the fit parameters.}

\item{title}{A string for the title of the plot}

\item{mean_width}{An integer of how many points to use for the means.}

\item{fit_width}{An integer of how many points to use for the fit.
If fit_width is smaller than mean_width then it takes all points.}

\item{manual_check}{A boolean to regulate if you want to manual check the data}
}
\value{
Returns a list of the check state, Ci1, Ci2, deltaCi, other constands
}
\description{
A function that fits Ci of a photosynthetic induction curve.
  It fits the data from the point where the light changes.
  It calculates Ci1, Ci2 and deltaCi.
  It uses nls to fit the data.
  You can check if the fit is good.
  It returns a list with all the parameters it used to fit the line.
  These include Ci1, Ci2, deltaCi and other constands.
  If the model gave a warning or error then the parameter values are NA.

  It only accepts licor files that have every input parameter stable except
  changing the light once. Commonly made with a timed lamp autoprogram with
  only two steps. The dataframe has to contain Ci, elapsed and Qin.
}
\examples{
  parameters <- fit_photosynthetic_induction_Ci_curve(dataframe,
                                                     "data from today")

}
\author{
Sam Loontjens
}
