% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide.R
\name{guide}
\alias{guide}
\title{A function that prints a guide of this package for the user.}
\usage{
guide()
}
\description{
A function that prints a guide of this package for the user.
}
\examples{
guide()

}
\author{
Sam Loontjens
}
