% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_load_packages.R
\name{install_load_packages}
\alias{install_load_packages}
\title{A function that installs packages.}
\usage{
install_load_packages(packages)
}
\arguments{
\item{packages}{A list of packages that have to be installed and loaded}
}
\value{
Nothing
}
\description{
A function that installs packages if they are not yet installed.
  Then it loads them with the library function.
}
\examples{
install_load_packages()

}
