% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_single_string.R
\name{is_single_string}
\alias{is_single_string}
\title{A function checks if the input is a single string.}
\usage{
is_single_string(input)
}
\arguments{
\item{input}{An object to check}
}
\value{
Returns a boolean
}
\description{
A function checks if the input is a single string.
  It returns a boolean.
}
\examples{
if (is_single_string(model)) {
  print("model gave an error or warning")
}

}
