% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_licorfiles.R
\name{list_licorfiles}
\alias{list_licorfiles}
\title{A function that lists the licorfiles in the input directory}
\usage{
list_licorfiles(pathname = "input_directory_licorfiles/")
}
\arguments{
\item{directory}{A directory name in the input directory
to read for licorfiles}
}
\value{
Returns a list of licorfile names in the given directory
}
\description{
A function that lists the licorfiles in the input directory.
  You can specify more directories to be more specific
}
\examples{
list_licorfiles("photosynthetic_induction_data")

}
