% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{moving_average}
\alias{moving_average}
\title{A function that calculates a moving average of a column or vector.}
\usage{
moving_average(column, averaging_width)
}
\arguments{
\item{column}{A column or vector to be averaged.}

\item{averaging_width}{the amount of neighbouring values to average.}
}
\value{
Returns the new column or vector that is moving averaged
to overwrite the old one or add as a new column.
}
\description{
A function that calculates a moving or rolling average of a
  column or vector.
  It returns the new column or vector that is averaged.
}
\examples{
mydata$A <- moving_average(mydata$A, 5)

or
mydata$avrA <- moving_average(mydata$A, 5)

}
