% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_dataframe.R
\name{numeric_dataframe}
\alias{numeric_dataframe}
\title{A function that changes all columns of a dataframe to numeric.}
\usage{
numeric_dataframe(dataframe)
}
\arguments{
\item{dataframe}{A dataframe to make numeric.}
}
\value{
Returns the new dataframe that is completely numeric
to overwrite the old one or add to a new dataframe.
}
\description{
A function that changes all columns of a dataframe to numeric.
  It returns the new dataframe that is completely numeric.
}
\examples{
mydata <- numeric_dataframe(mydata)

}
\author{
Sam Loontjens
}
