% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_licorfile.R
\name{read_licorfile}
\alias{read_licorfile}
\title{A function that reads a licorfile.}
\usage{
read_licorfile(filename, sheetnumber = 1, parameters = TRUE)
}
\arguments{
\item{filename}{A string with the filename to read.}

\item{sheetnumber}{An integer which sheet to read from the file.}

\item{parameters}{A list of desired parameters to extract.}
}
\value{
Returns a dataframe with only the desired parameters.
}
\description{
A function that reads a licorfile.
  It returns a clean dataframe with only the desired parametes.
}
\examples{
parameters_to_extract <- c("A, Ci, elapsed, Qin")
mydata <- read_licorfile("20210226 PI 50-100 75RH 400CO2 T.xlsx",
                         parameters_to_extract)

}
\author{
Sam Loontjens
}
