% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_light.R
\name{round_light}
\alias{round_light}
\title{A function that rounds the Qin of a dataframe.}
\usage{
round_light(dataframe)
}
\arguments{
\item{dataframe}{A dataframe to round the Qin.}
}
\value{
Returns a dataframe with the rounded Qin
to overwrite the old one or to add a new dataframe.
}
\description{
A function that rounds the Qin of a dataframe.
  It returns a dataframe with the rounded Qin.
}
\examples{
mydata <- round_light(mydata)

}
\author{
Sam Loontjens
}
