% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_parameter.R
\name{round_parameter}
\alias{round_parameter}
\title{A function that rounds one column or vector of a dataframe.}
\usage{
round_parameter(dataframe, parameter)
}
\arguments{
\item{dataframe}{A dataframe}

\item{parameter}{A column name that is the parameter to round}
}
\value{
Returns a new column or vector that is rounded.
}
\description{
A function that rounds one column or vector of a dataframe.
  It returns a new column or vector that is rounded.
}
\examples{
mydata$Qin <- round_parameter(mydata$Qin)

}
