% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hi.R
\name{say_hi}
\alias{say_hi}
\title{A function to say hi to Sam.}
\usage{
say_hi(snack = "cookie")
}
\arguments{
\item{snack}{A string that describes a nice snack}
}
\value{
Nothing
}
\description{
A function that says hi to Sam and asks for a cookie.
}
\examples{
say_hi("Banana")

}
\author{
Sam Loontjens
}
