% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_filepath.R
\name{split_filepath}
\alias{split_filepath}
\title{A function that splits a filepath.}
\usage{
split_filepath(pathname)
}
\arguments{
\item{pathname}{A pathname to split}
}
\value{
A list with directory names and a filename at the end
}
\description{
A function that splits a filepath.
  It returns a list of the directories and the filename.
}
\examples{
split_filepath("/input_directory_licorfiles/photosynthetic_induction_data/
               20210226 PI 50-100 75RH 400CO2 T.xlsx")

}
\author{
Sam Loontjens
}
