% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_licor_filename.R
\name{split_licor_filename}
\alias{split_licor_filename}
\title{A function that splits a filename of a licorfile.}
\usage{
split_licor_filename(filename)
}
\arguments{
\item{filename}{A filename to split}
}
\value{
A list of the parameters included in the filename.
}
\description{
A function that splits a filename of a licorfile.
  It returns a list of the parameters included in the filename.
}
\examples{
split_filepath("20210226 PI 50-100 75RH 400CO2 T.xlsx")

}
\author{
Sam Loontjens
}
