% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{startup}
\alias{startup}
\title{A function that runs the reccomended functions before starting analysis.}
\usage{
startup()
}
\value{
Returns TRUE if the startup was completed
}
\description{
A function that runs the required functions before starting analysis.
  It downloads and loads the required packages
  It sets the working directory to the path of the source file
  It creates directories where you can drop files for analysis
  Prints "startup complete" when finished
}
\examples{
startup()

}
