% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_photosynthetic_induction_data.R
\name{summarise_photosynthetic_induction_data}
\alias{summarise_photosynthetic_induction_data}
\title{A function that combines and summarises photosynthetic induction data.}
\usage{
summarise_photosynthetic_induction_data(
  manual_check = TRUE,
  write_excel = TRUE,
  summarise = TRUE,
  save_plot = TRUE
)
}
\arguments{
\item{manual_check}{A boolean that regulates if the fits are checked.
Default is TRUE.}

\item{write_excel}{A boolean that regulates if the dataframe is written to an excel file.
Default is TRUE.}

\item{summarise}{A boolean that regulates if the data is summarised.
Default is TRUE.}

\item{save_plot}{A boolean that handles if the plot get saved.
Default is TRUE.}
}
\value{
Returns a list of dataframes with the combined data and summarised data.
}
\description{
##NOTE: Not completed and tested yet. Needs to be updated for more functionality##
  A function that combines the desired parameters of multiple photosynthetic
  induction licorfiles and summarises them to calculate the mean and standard
  error for a plot.
}
\examples{
mydata <- summarise_photosynthetic_induction_data()

}
\author{
Sam Loontjens
}
