% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_dataframe_names.R
\name{unique_dataframe_names}
\alias{unique_dataframe_names}
\title{A function that renames duplicate dataframe column names}
\usage{
unique_dataframe_names(dataframe)
}
\arguments{
\item{dataframe}{A dataframe}
}
\value{
Returns a dataframe with unique names
}
\description{
A function that renames duplicate dataframe column names
  Makes it so the database can be used for ggplot which requires
  all columns to have a unique name
}
\examples{

}
\author{
Sam Loontjens
}
